pyproject.toml
[tool.poetry]
name = "reddit-fetcher"
version = "0.1.0"
description = "A script to fetch the latest posts from a subreddit using PRAW."
authors = ["Natasha Sharma <nssharma.official@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
asyncpraw = "^7.7.1"
python-dotenv = "^1.0.1"
aiohttp = "^3.9.1"
aiolimiter = "^1.1.0"
fastapi = "^0.110.0"
uvicorn = "^0.27.1"
pydantic = "^2.6.3"
openai = "^1.1.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api" 